more fixes for imports
