+++ src/consts.py
@@ -11,15 +11,6 @@ class SUBSCRIPTIONS(enum.Enum):
     TROVE = 'Humble Trove'
 
 
-class KEY_TYPE(enum.Enum):
-    STEAM = 'steam'
-    ORIGIN = 'origin'
-    UPLAY = 'uplay'
-    EPIC = 'epic'  # not sure about it
-    BATTLENET = 'battlenet'  # not sure about it
-    GOG = 'gog'  # not sure about it
-
-
 class SOURCE(enum.Enum):
     DRM_FREE = 'drm-free'
     KEYS = 'keys'
@@ -29,8 +20,6 @@ class BITNESS(enum.Enum):
     B64 = 64
     B32 = 32
 
-
-GAME_PLATFORMS = set([HP.WINDOWS, HP.MAC, HP.LINUX])
 NON_GAME_BUNDLE_TYPES = {'mobilebundle', 'softwarebundle', 'bookbundle', 'audiobookbundle', 'comicsbundle', 'rpgbookbundle', 'mangabundle'}
 
 IS_WINDOWS = sys.platform == 'win32'
